# docs/openapi.yaml
openapi: 3.1.0
info:
  title: La Trobe Content Distribution API
  version: 3.0.0
  description: Assessment 3 publishing, RSS, operational metrics, status, and alert contract.
servers:
  - url: http://localhost:4000
...
paths:
...
  /rss/{channelCode}:
    get:
      tags: [RSS]
      summary: Read one Channel as RSS 2.0
      parameters:
        - {
            name: channelCode,
            in: path,
            required: true,
            schema: { type: string, example: CLOUDDEVOPS },
          }
        - { $ref: "#/components/parameters/ClientId" }
        - { $ref: "#/components/parameters/ClientSource" }
      responses:
        "200": { description: Channel RSS feed, content: { application/rss+xml: {} } }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  parameters:
...
    ClientId:
      name: X-Client-Id
      in: header
      schema: { type: string, maxLength: 120 }
    ClientSource:
      name: X-Client-Source
      in: header
      schema: { type: string, maxLength: 80 }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
  responses:
...
    NotFound:
      description: Record was not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }